backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
beautifulsoup4
requests
python-dotenv
transformers
torch
groq
psycopg2-binary
slowapi
